refactor(embedded): restore tmpl layout and drop tmplgen package
Move the template generator back under embedded/tmpl/ with the original
main_test.go (//go:embed gen). Consolidate Bazel targets into
embedded/BUILD.bazel as on main and remove the tmplgen subdirectory.
