+++ pkg/fleet/installer/packages/embedded/tmpl/datadog-agent-ddot.yaml.tmpl
