+++ pkg/fleet/installer/packages/embedded/tmpl/datadog-agent.service.tmpl
