+++ pkg/fleet/installer/packages/embedded/tmpl/main_test.go
@@ -1,0 +1,43 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package main generates the systemd units for the installer.
+package main
+
+import (
+	"embed"
+	"io/fs"
+	"os"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/fixtures"
+)
+
+//go:embed gen
+var genFS embed.FS
+
+// TestGenerationIsUpToDate tests that the generated templates are up to date.
+//
+// You can update the templates by running `go generate` in the templates directory.
+func TestGenerationIsUpToDate(t *testing.T) {
+	if os.Getenv("CI") == "true" && runtime.GOOS == "darwin" {
+		t.Skip("TestGenerationIsUpToDate is known to fail on the macOS Gitlab runners.")
+	}
+
+	generated := filepath.Join(os.TempDir(), "gen")
+	os.MkdirAll(generated, 0755)
+
+	err := generate(generated)
+	assert.NoError(t, err)
+	newGeneratedFS := os.DirFS(generated)
+	currentGeneratedFS, err := fs.Sub(genFS, "gen")
+	assert.NoError(t, err)
+
+	fixtures.AssertEqualFS(t, currentGeneratedFS, newGeneratedFS)
+}
