--- pkg/fleet/installer/packages/embedded/tmplgen/BUILD.bazel
@@ -1,45 +1,0 @@
-load("@rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
-
-exports_files(
-    glob(["*.tmpl"]),
-    visibility = ["//pkg/procmgr/rust:__pkg__"],
-)
-
-go_library(
-    name = "tmplgen_lib",
-    srcs = ["main.go"],
-    embedsrcs = [
-        "datadog-agent-action.service.tmpl",
-        "datadog-agent-data-plane.service.tmpl",
-        "datadog-agent-ddot.service.tmpl",
-        "datadog-agent-ddot.yaml.tmpl",
-        "datadog-agent-installer.service.tmpl",
-        "datadog-agent-process.service.tmpl",
-        "datadog-agent-procmgr.service.tmpl",
-        "datadog-agent-security.service.tmpl",
-        "datadog-agent-sysprobe.service.tmpl",
-        "datadog-agent-trace.service.tmpl",
-        "datadog-agent.service.tmpl",
-    ],
-    importpath = "github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/embedded/tmplgen",
-    visibility = ["//visibility:private"],
-    deps = ["//pkg/template/text"],
-)
-
-go_binary(
-    name = "tmplgen",
-    embed = [":tmplgen_lib"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "tmplgen_test",
-    srcs = ["main_test.go"],
-    data = ["//pkg/fleet/installer/packages/embedded:tmpl_gen"],
-    embed = [":tmplgen_lib"],
-    gotags = ["test"],
-    deps = [
-        "//pkg/fleet/installer/fixtures",
-        "@com_github_stretchr_testify//require",
-    ],
-)
